Comment out the check in mlx5_modify_qp_rate_limit() on the MLX5_IB_PP_SUPPORT_BURST flag, to allow the request to be passed to the kernel.
This is because MLX5_IB_PP_SUPPORT_BURST is only set when both max_burst_sz and typical_pkt_sz can be modified.

Whereas the kernel allows just max_burst_sz to be modified even if the device doesn't also allow typical_pkt_sz to be modified (as appears to be the case with a ConnectX-4 Lx).